% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_point_process.R
\name{simulate_point_process}
\alias{simulate_point_process}
\title{Simulate a spatial pattern}
\usage{
simulate_point_process(
  spatial_association,
  gamma,
  radius,
  num_candidates = 1000,
  Nsteps = 50
)
}
\arguments{
\item{spatial_association}{a matrix with elements encoding the strength of species associations. Strength in (0, 1) indicates attraction, = 0 indicates null relationship, in (-1, 0) indicates repulsion}

\item{gamma}{the gamma diversity}

\item{radius}{the range of effective size}

\item{num_candidates}{the number of candidate points that the sample are from}

\item{Nsteps}{the number of total simulated points}
}
\description{
Simulate a spatial pattern
}
