% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_association.R
\name{plot_spatial_association}
\alias{plot_spatial_association}
\title{Plot the spatial associations}
\usage{
plot_spatial_association(spatial_association)
}
\arguments{
\item{spatial_association}{a matrix with elements encoding the strength of species associations. Strength in (0, 1) indicates attraction, = 0 indicates null relationship, in (-1, 0) indicates repulsion}
}
\value{
A graph visualizing spatial associations
}
\description{
Plot the spatial associations
}
