% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_association.R
\name{plot_spatial_association}
\alias{plot_spatial_association}
\title{Plot the spatial associations}
\usage{
plot_spatial_association(spatial_association, plot_type = "ggraph")
}
\value{
A graph visualizing spatial associations
}
\description{
Plot the spatial associations
}
