% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spatial_association.R
\name{sample_spatial_point}
\alias{sample_spatial_point}
\title{Generate a new spatial location for a given species}
\usage{
sample_spatial_point(
  species_location,
  spatial_association,
  gamma,
  invading_species,
  num_candidates,
  radius
)
}
\description{
Generate a new spatial location for a given species
}
